replace occurences of VERSION with centerim::version
This patch adds a new file git-version.cc, which intializes static variable
centerim::version.

This helps in cutting down the compile time penalty of generating the
version string on every make run. Instead of every file that uses config.h
only one file has to be recompiled. The rest of the work is done by the
linker.