Adding interpretation and saving of server capabilities. Commented out sending CL at login.
Formerly the server capabilities were just ignored. Now they are saved in a
vector so that we may re-use them to prevent the sending of unrecognized
commands.
Also commented out the sending of the Contact List to server at login time
since it is not part of the login steps and the server responds with an error.