Fixed compiler error with FreeBSD. Also fixed link error against curseslib.
FreeBSD do not want <utmp.h> to be included. However utmp is only used in linux,
so we do not need to include utmp.h for non linux systems.

Also bug in configure script was fixed.
The curses-lib was not properly included in the $LIBS-variable in the
configure.ac file.